Fixed dll copying for indirect references
Previously, the dlls would get copied to the output directory of the
project referencing the nuget package, but not for projects referencing
a project that references the nuget package. So, if you have an app that
has a library that uses DirectXTex.net, it won't get the right dlls and
will crash at runtime.

I changed it to hook into the existing copy content msbuild machinery
rather than copying manually, which handles this recursive copying, and
is actually a bit simpler.